Conflicts:
	apps/addonsExamples/3DModelLoaderExample/config.make
	apps/addonsExamples/allAddonsExample/config.make
	apps/addonsExamples/networkTcpClientExample/config.make
	apps/addonsExamples/networkTcpServerExample/config.make
	apps/addonsExamples/networkUdpReceiverExample/Makefile
	apps/addonsExamples/networkUdpReceiverExample/config.make
	apps/addonsExamples/networkUdpSenderExample/Makefile
	apps/addonsExamples/networkUdpSenderExample/config.make
	apps/addonsExamples/ofxAssimpExample/Makefile
	apps/addonsExamples/ofxAssimpExample/config.make
	apps/addonsExamples/ofxCvHaarFinderExample/Makefile
	apps/addonsExamples/ofxCvHaarFinderExample/config.make
	apps/addonsExamples/ofxSynthExample/Makefile
	apps/addonsExamples/ofxSynthExample/config.make
	apps/addonsExamples/ofxSynthSequencingExample/Makefile
	apps/addonsExamples/ofxSynthSequencingExample/config.make
	apps/addonsExamples/opencvExample/Makefile
	apps/addonsExamples/opencvExample/config.make
	apps/addonsExamples/oscReceiveExample/Makefile
	apps/addonsExamples/oscReceiveExample/config.make
	apps/addonsExamples/oscSenderExample/Makefile
	apps/addonsExamples/oscSenderExample/config.make
	apps/addonsExamples/threadedImageLoaderExample/Makefile
	apps/addonsExamples/threadedImageLoaderExample/config.make
	apps/addonsExamples/vectorGraphicsExample/Makefile
	apps/addonsExamples/vectorGraphicsExample/config.make
	apps/addonsExamples/xmlSettingsExample/Makefile
	apps/addonsExamples/xmlSettingsExample/config.make
	apps/advancedExamples/geometryShaderExample/Makefile
	apps/advancedExamples/threadExample/Makefile
	apps/androidExamples/androidAssimpExample/Makefile
	apps/androidExamples/androidAssimpExample/config.make
	apps/androidExamples/androidAudioExample/Makefile
	apps/androidExamples/androidAudioExample/config.make
	apps/androidExamples/androidCameraExample/Makefile
	apps/androidExamples/androidCameraExample/config.make
	apps/androidExamples/androidEmptyExample/Makefile
	apps/androidExamples/androidEmptyExample/config.make
	apps/androidExamples/androidFontExample/Makefile
	apps/androidExamples/androidFontExample/config.make
	apps/androidExamples/androidImageExample/Makefile
	apps/androidExamples/androidImageExample/config.make
	apps/androidExamples/androidOpenCVExample/Makefile
	apps/androidExamples/androidOpenCVExample/config.make
	apps/androidExamples/androidPolygonExample/Makefile
	apps/androidExamples/androidPolygonExample/config.make
	apps/androidExamples/androidSoundPlayerExample/Makefile
	apps/androidExamples/androidSynthExample/Makefile
	apps/androidExamples/androidSynthExample/config.make
	apps/androidExamples/androidTouchExample/Makefile
	apps/androidExamples/androidTouchExample/config.make
	apps/androidExamples/androidVBOExample/Makefile
	apps/androidExamples/androidVBOExample/config.make
	apps/devApps/linux/Makefile
	apps/devApps/linux/config.make
	apps/devApps/synthExample/Makefile
	apps/devApps/synthSequencingExample/Makefile
	apps/examples/advanced3dExample/Makefile
	apps/examples/advanced3dExample/config.make
	apps/examples/advancedEventsExample/Makefile
	apps/examples/advancedEventsExample/config.make
	apps/examples/audioInputExample/Makefile
	apps/examples/audioInputExample/config.make
	apps/examples/audioOutputExample/Makefile
	apps/examples/audioOutputExample/config.make
	apps/examples/blendingExample/Makefile
	apps/examples/blendingExample/config.make
	apps/examples/dirListExample/Makefile
	apps/examples/dirListExample/config.make
	apps/examples/dragDropExample/Makefile
	apps/examples/dragDropExample/config.make
	apps/examples/easyCamExample/Makefile
	apps/examples/easyCamExample/config.make
	apps/examples/emptyExample/Makefile
	apps/examples/emptyExample/config.make
	apps/examples/eventsExample/Makefile
	apps/examples/eventsExample/config.make
	apps/examples/firmataExample/Makefile
	apps/examples/firmataExample/config.make
	apps/examples/fontShapesExample/Makefile
	apps/examples/fontShapesExample/config.make
	apps/examples/fontsExample/Makefile
	apps/examples/fontsExample/config.make
	apps/examples/graphicsExample/Makefile
	apps/examples/graphicsExample/config.make
	apps/examples/imageLoaderExample/Makefile
	apps/examples/imageLoaderExample/config.make
	apps/examples/imageSaverExample/Makefile
	apps/examples/imageSaverExample/config.make
	apps/examples/movieGrabberExample/Makefile
	apps/examples/movieGrabberExample/config.make
	apps/examples/moviePlayerExample/Makefile
	apps/examples/moviePlayerExample/config.make
	apps/examples/ofBoxExample/Makefile
	apps/examples/ofBoxExample/config.make
	apps/examples/polygonExample/Makefile
	apps/examples/polygonExample/config.make
	apps/examples/serialExample/Makefile
	apps/examples/serialExample/config.make
	apps/examples/soundPlayerExample/Makefile
	apps/examples/soundPlayerExample/config.make
	apps/examples/soundPlayerFFTExample/Makefile
	apps/examples/soundPlayerFFTExample/config.make
	apps/examples/textureExample/Makefile
	apps/examples/textureExample/config.make
	apps/examples/textureScreengrabExample/Makefile
	apps/examples/textureScreengrabExample/config.make
	apps/examples/vectorMathExample/Makefile
	apps/examples/vectorMathExample/config.make
	apps/examples/windowExample/Makefile
	apps/examples/windowExample/config.make
	scripts/linux/template/Makefile

+++ scripts/linux/createProjects.py
@@ -1,4 +1,4 @@
-#!/usr/bin/python
+#!/usr/bin/env python
 import os
 from lxml import etree
 from lxml import objectify
@@ -6,11 +6,17 @@ import argparse
 import shutil
 import glob
 
-of_root = '/home/arturo/Desktop/openFrameworks'
+
+of_root = os.path.realpath(__file__)[0:-(len(os.path.join('scripts','linux','createProject.py'))+2)]
 platform = 'linux'
-arch = 'linux64'
-templates_path = '../../apps/examples/emptyExample/'
-template = {'cbp': templates_path + 'emptyExample_' + arch + '.cbp', 'makefile': templates_path + 'Makefile', 'config.make': templates_path + 'config.make'}
+
+arch = 'linux'
+
+uname = os.uname()
+for uname_str in uname:
+    if uname_str=='x86_64':
+        arch = 'linux64'
+
 fullCBP = True
 
 def addCBPIncludePath(project,dirpath):
@@ -27,7 +33,6 @@ def addCBPIncludePath(project,dirpath):
         include.set("directory",dirpath)
         
 def addCBPLibrary(project,libpath):
-    print libpath
     found=False
     if project.find('Linker') == None:
         etree.SubElement(project,"Linker")
@@ -83,14 +88,15 @@ def addAddon(project,addon):
                 continue
             basefolder = os.path.join('addons',addon,'libs',libdir);
             if os.path.exists(os.path.join(of_root,basefolder,'include')):
-                dirpath = os.path.join('..','..','..',basefolder,'include')
-                addCBPIncludePath(project,dirpath)
-                for root, dirs, files in os.walk(os.path.join(libdir,'include')):
+                dirpath = os.path.join(of_root,basefolder,'include')
+                addCBPIncludePath(project,os.path.join('..','..','..',basefolder,'include'))
+                for root, dirs, files in os.walk(dirpath):
                     for dir in dirs:
-                        basefolder = root[len(of_root)+1:]
-                        dirpath = os.path.join('..','..','..',basefolder,dir)
-                        addCBPIncludePath(project,dirpath)
+                        basefolder_addon = root[len(of_root)+1:]
+                        dirpath_addon = os.path.join('..','..','..',basefolder_addon,dir)
+                        addCBPIncludePath(project,dirpath_addon)
             
+            basefolder = os.path.join('addons',addon,'libs',libdir);
             if os.path.exists(os.path.join(of_root,basefolder,'lib',arch)):
                 dirpath = os.path.join('..','..','..',basefolder,'lib',arch)
                 if os.path.exists(os.path.join(of_root,basefolder,'lib',arch,'libsorder.make')):
@@ -104,11 +110,9 @@ def addAddon(project,addon):
                     for lib in glob.glob(os.path.join(of_root,basefolder,'lib',arch,'*.a')):
                         baselib = lib[len(of_root)+1:]
                         addCBPLibrary(project,os.path.join('..','..','..',baselib))
-                        print baselib
                     for lib in glob.glob(os.path.join(of_root,basefolder,'lib',arch,'*.so')):
                         baselib = lib[len(of_root)+1:]
                         addCBPLibrary(project,os.path.join('..','..','..',baselib))
-                        print baselib
                         
 
 def addAddons(project,project_path):
@@ -145,8 +149,28 @@ def createCBP(project_path):
     cbp_file = open(os.path.join(project_path,project_name+'.cbp'),mode='w')
     cbp_file.write(etree.tostring(cbp, xml_declaration=True, encoding='UTF-8', pretty_print=True))
     cbp_file.close()
+    
+def createWorkspace(project_path):
+    if os.path.abspath(project_path) == os.path.abspath(templates_path):
+        return
+    project_name = os.path.basename(project_path)
+    ws = objectify.parse(os.path.join(project_path,project_name+'.workspace'))
+    root = ws.getroot()
+    workspace = root.Workspace
+
+    if workspace.get("title")=="emptyExample":
+         workspace.set("title",project_name)
+         
+    for project in workspace.Project:
+        if project.get("filename")=="emptyExample.cbp":
+            project.set("filename",project_name+".cbp")
+
+    ws_file = open(os.path.join(project_path,project_name+'.workspace'),mode='w')
+    ws_file.write(etree.tostring(ws, xml_declaration=True, encoding='UTF-8', pretty_print=True))
+    ws_file.close()
 
 def createProject(project_path):
+    print 'generating',project_path
     if os.path.abspath(project_path) == os.path.abspath(templates_path):
         return
     if project_path[-1]==os.sep:
@@ -155,10 +179,17 @@ def createProject(project_path):
         os.mkdir(project_path)
     
     project_name = os.path.basename(project_path)
-    shutil.copyfile(template['cbp'],os.path.join(project_path,project_name+'.cbp'))
-    shutil.copyfile(template['makefile'],os.path.join(project_path,'Makefile'))
+    if fullCBP:
+        shutil.copyfile(template['full_cbp'],os.path.join(project_path,project_name+'.cbp'))
+    else:
+        shutil.copyfile(template['cbp'],os.path.join(project_path,project_name+'.cbp'))
+        
+    shutil.copyfile(template['workspace'],os.path.join(project_path,project_name+'.workspace'))
+    
+    if platform == "linux":
+        shutil.copyfile(template['makefile'],os.path.join(project_path,'Makefile'))
 
-    if not os.path.exists(os.path.join(project_path, 'config.make')):
+    if platform == "linux" and not os.path.exists(os.path.join(project_path, 'config.make')):
         shutil.copyfile(template['config.make'],os.path.join(project_path,'config.make'))
 
     if not os.path.exists(os.path.join(project_path,'src')):
@@ -173,21 +204,28 @@ def createProject(project_path):
         os.mkdir(os.path.join(project_path , 'bin','data'))
 
     createCBP(project_path)
+    createWorkspace(project_path)
 
 
 
 parser = argparse.ArgumentParser(description='OF linux project generator')
 parser.add_argument('project_path', metavar='project_path', nargs='?')
+parser.add_argument('-n', '--not_mk', dest='not_mk', action='store_const',
+        default=False, const=True, help='create cbp not dependent on Makefile')
+parser.add_argument('-p', '--platform', dest='platform', action='store', default=arch, choices=['linux','linux64'], help='choose platform: linux/linux64 if it\'s not specified it\'ll be detected')
+
+arch = parser.parse_args().platform
 project_path = parser.parse_args().project_path
+fullCBP = parser.parse_args().not_mk
+        
+templates_path = os.path.join(of_root,'scripts',platform,'template')
+template = {'cbp': os.path.join(templates_path , 'emptyExample_' + arch + '.cbp'), 'full_cbp': os.path.join(templates_path , 'emptyExample_' + arch + '_fullCBP.cbp'), 'workspace': os.path.join(templates_path , 'emptyExample_' + arch + '.workspace'),'makefile': os.path.join(templates_path , 'Makefile'), 'config.make': os.path.join(templates_path , 'config.make')}
 
 if project_path==None: #parse all examples
-    #for root, dirs, files in os.walk('../../apps/examples'):
-    #    for name in dirs:
-    #        print os.path.join(root,name)
-    for example in os.listdir(os.path.join('..','..','apps','examples')):
-        createProject(os.path.join('..','..','apps','examples',example))
-    for example in os.listdir(os.path.join('..','..','apps','addonsExamples')):
-        createProject(os.path.join('..','..','apps','addonsExamples',example))
+    for example in os.listdir(os.path.join(of_root,'apps','examples')):
+        createProject(os.path.join(of_root,'apps','examples',example))
+    for example in os.listdir(os.path.join(of_root,'apps','addonsExamples')):
+        createProject(os.path.join(of_root,'apps','addonsExamples',example))
 else:
     createProject(project_path)
